Changes made to misc05 and loader files

+++ common/objloader.hpp
@@ -11,7 +11,7 @@ bool loadOBJ(
 
 bool loadAssImp(
 	const char * path, 
-	std::vector<unsigned short> & indices,
+	std::vector<unsigned int> & indices,
 	std::vector<glm::vec3> & vertices,
 	std::vector<glm::vec3> & normals
 );
