+++ common/vboindexer.hpp
@@ -5,7 +5,7 @@ void indexVBO(
 	std::vector<glm::vec3> & in_vertices,
 	std::vector<glm::vec3> & in_normals,
 
-	std::vector<unsigned short> & out_indices,
+	std::vector<unsigned int> & out_indices,
 	std::vector<glm::vec3> & out_vertices,
 	std::vector<glm::vec3> & out_normals
 );
